Move JDK11-specific Zip inflate/deflater stuff to JDK11 module.
Fixes #1423.
